First pass linting for .es1 dictionary
The dictionaries are designed to work as standalone source, and include
numbers, special characters, some random letter-number combination that
have no meaning, and various spurious stuff.

This is a first pass at linting the .es1 dictionary, doesn't include a
re-indexing, just weeding out part of what is not Spanish words.